+++ docs/constants/navigation.js
@@ -308,6 +308,7 @@ const general = [
     makeGroup('Reference', [
       makePage('eas-update/migrate-from-classic-updates.mdx'),
       makePage('eas-update/codepush.mdx'),
+      makePage('eas-update/updating-your-app.mdx'),
       makePage('eas-update/faq.mdx'),
       makePage('eas-update/known-issues.mdx'),
     ]),
@@ -353,7 +354,6 @@ const general = [
       makePage('workflow/upgrading-expo-sdk-walkthrough.mdx'),
       makePage('eas/webhooks.mdx'),
       makeSection('Assorted', [
-        makePage('guides/routing-and-navigation.mdx'),
         makePage('guides/authentication.mdx'),
         makePage('guides/troubleshooting-proxies.mdx'),
         makePage('guides/sharing-preview-releases.mdx'),
@@ -385,7 +385,6 @@ const general = [
         makePage('bare/using-expo-cli.mdx'),
         makePage('bare/installing-updates.mdx'),
         makePage('bare/using-expo-client.mdx'),
-        makePage('bare/updating-your-app.mdx'),
         makePage('bare/install-dev-builds-in-bare.mdx'),
         makePage('bare/error-recovery.mdx'),
       ]),